requirements.txt
alembic==1.4.2
Babel==2.8.0
blinker==1.4
click==7.1.2
dominate==2.5.1
Flask==1.1.2
Flask-Babel==1.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.2
MarkupSafe==1.1.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
six==1.14.0
SQLAlchemy==1.3.17
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.1
